feat: Implement core functionality for Chimauta application
- Add main application class `ChimautaApp` to manage initialization and state.
- Create `DataManager` for fetching and processing data from the API.
- Introduce `DOMElements` class for managing DOM elements.
- Implement `YouTubePlayerManager` to handle YouTube player interactions.
- Add `SearchManager` for search functionality and URL parameter handling.
- Create `AppState` class to manage application state.
- Develop `UIManager` for user interface interactions and updates.
- Define constants in `constants.js` for API URLs and configuration.
- Structure the application with modular JavaScript files for better maintainability.
