+++ frontend/index.html
@@ -43,7 +43,15 @@
 
     <!-- My JS -->
     <script defer src="js/util.js"></script>
-    <script defer src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
+    <script type="module" src="js/app.js"></script>
+    <script type="module" src="js/constants.js"></script>
+    <script type="module" src="js/data.js"></script>
+    <script type="module" src="js/dom.js"></script>
+    <script type="module" src="js/player.js"></script>
+    <script type="module" src="js/search.js"></script>
+    <script type="module" src="js/state.js"></script>
+    <script type="module" src="js/ui.js"></script>
 
     <!--Font Awesome-->
     <script defer src="https://kit.fontawesome.com/ddd08aa620.js" crossorigin="anonymous"></script>
