+++ frontend/js/constants.js
@@ -1,0 +1,13 @@
+const CONSTANTS = {
+  API_URL:
+    "https://script.google.com/macros/s/AKfycbzANnxCbUliOy7eDUuGFBPRraKOKHvq_2-SsxGQMlZ10MaLEOAJZr6G9uawYieRK2ei/exec",
+  YOUTUBE_API_URL: "https://www.youtube.com/iframe_api",
+  UPDATE_INTERVAL: 250,
+  ERROR_RETRY_DELAY: 5000,
+  THUMB_BASE_URL: "https://img.youtube.com/vi_webp/",
+  YOUTUBE_WATCH_URL: "https://youtu.be/",
+  TWITTER_INTENT_URL:
+    "https://x.com/intent/tweet?ref_src=twsrc%5Etfw%7Ctwcamp%5Ebuttonembed%7Ctwterm%5Eshare%7Ctwgr%5E",
+};
+
+export default CONSTANTS;
